+++ backend/garpix_notify/serializers.py
@@ -1,0 +1,10 @@
+from rest_framework import serializers
+
+from garpix_notify.models import SystemNotify
+
+
+class SystemNotifySerializer(serializers.ModelSerializer):
+
+    class Meta:
+        model = SystemNotify
+        fields = '__all__'